run: Be able to package before running instead of just building
